requirements.txt
duckduckgo-search>=4.1.1
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
tabulate>=0.9.0
pydantic>=2.0.0
groq>=0.4.2
aiohttp>=3.8.0 